Put log stream in it's own file, makes it easier to reuse.
..and test. I had issue with logging out of order
on Heroku and I though it was something fishy going
on with my writeable stream. Seems that it is ok,
and tested locally logging is written in order.

On heroku however it is out of order, and according
to their docs https://devcenter.heroku.com/articles/logging#log-message-ordering
this may be the case..