Detect the correct content type based on the first 16k in the stream (#11)
The URL to be uploaded might return `binary/octet-stream` as the content-type,
which is not particularly useful as the type to store in S3. Luckily there is a
neat little package we can use that reliably detects the type without having to
load the whole file into memory.

Although `stream-mmmagic` supports returning a promise, I found it a lot easier
to use the callback style to ensure we handle any errors correctly.