Allow `makePublic`being `null`
In #10 support was added to conditionally making the files public by providing a boolean property `makePublic`. I’m seeing an increased number of incidents now where the requests fails due to this prop being `null`, which is not valid in JSON Schema.

Instead of fixing our many legacy clients I’m instead just patching this service to accept `null` as a value.